+++ src/plugins/cmakeprojectmanager/builddirmanager.cpp
@@ -280,7 +280,7 @@ void BuildDirManager::stopProcess()
 
     if (m_cmakeProcess->state() == QProcess::Running) {
         m_cmakeProcess->terminate();
-        if (!m_cmakeProcess->waitForFinished(500))
+        if (!m_cmakeProcess->waitForFinished(500) && m_cmakeProcess->state() == QProcess::Running)
             m_cmakeProcess->kill();
     }
 
@@ -305,9 +305,10 @@ void BuildDirManager::cleanUpProcess()
 
     if (m_cmakeProcess->state() == QProcess::Running) {
         m_cmakeProcess->terminate();
-        if (!m_cmakeProcess->waitForFinished(500))
+        if (!m_cmakeProcess->waitForFinished(500) && m_cmakeProcess->state() == QProcess::Running)
             m_cmakeProcess->kill();
     }
+    m_cmakeProcess->waitForFinished();
     delete m_cmakeProcess;
     m_cmakeProcess = nullptr;
